@@ -1,6 +1,6 @@
 {
   "name": "shallow-render",
-  "version": "8.5.2",
+  "version": "9.0.0-0",
   "description": "Shallow rendering test utility for Angular",
   "main": "dist/index.js",
   "files": [
@@ -17,17 +17,20 @@
   },
   "scripts": {
     "clean": "rm -rf dist coverage",
-    "lint": "tslint -t stylish --project tsconfig.json",
-    "lint:fix": "npm run lint -- --fix",
-    "build": "npm run clean && tsc",
+    "build": "npm run clean && tsc --project tsconfig.prod.json",
     "build:all": "npm run lint && npm run test && npm run build",
-    "preversion": "npm run build:all",
-    "postversion": "echo 'Now run:\ngit push --follow-tags && npm publish'",
     "test": "karma start",
     "test:fix": "npm run lint:fix && npm run test",
     "test:watch": "npm test -- --auto-watch=true --single-run=false",
     "test:debug": "npm run test:watch -- --browsers=Chrome",
-    "test:angular-versions": "./build-with-supported-angluars.js"
+    "lint": "npm run lint:ts && npm run lint:format",
+    "lint:ts": "tslint -t stylish --project tsconfig.json",
+    "lint:ts:fix": "npm run lint:ts -- --fix",
+    "lint:format": "prettier --list-different '**/*'",
+    "lint:format:fix": "prettier --write '**/*'",
+    "lint:fix": "npm run lint:ts:fix && npm run lint:format:fix",
+    "preversion": "npm run build:all",
+    "postversion": "echo 'Now run:\ngit push --follow-tags && npm publish'"
   },
   "repository": {
     "type": "git",
@@ -47,37 +50,41 @@
   },
   "homepage": "https://github.com/getsaf/shallow-render#readme",
   "peerDependencies": {
-    "@angular/common": ">=6.x <=8.x",
-    "@angular/compiler": ">=6.x <=8.x",
-    "@angular/core": ">=6.x <=8.x",
-    "@angular/forms": ">=6.x <=8.x",
-    "@angular/platform-browser": ">=6.x <=8.x"
+    "@angular/common": "9.x",
+    "@angular/compiler": "9.x",
+    "@angular/core": "9.x",
+    "@angular/forms": "9.x",
+    "@angular/platform-browser": "9.x"
   },
   "dependencies": {
-    "ng-mocks": "^7.8.0"
+    "ng-mocks": "^9.0.0"
   },
   "devDependencies": {
-    "@angular/common": "^6.1.10",
-    "@angular/compiler": "^6.1.10",
-    "@angular/core": "^6.1.10",
-    "@angular/forms": "^6.1.10",
-    "@angular/platform-browser": "^6.1.10",
-    "@angular/platform-browser-dynamic": "^6.1.10",
-    "@angular/router": "^6.1.10",
-    "@types/jasmine": "^2.5.53",
-    "@types/node": "^8.10.25",
+    "@angular/common": "9.x",
+    "@angular/compiler": "9.x",
+    "@angular/core": "9.x",
+    "@angular/forms": "9.x",
+    "@angular/platform-browser": "9.x",
+    "@angular/platform-browser-dynamic": "9.x",
+    "@angular/router": "9.x",
+    "@types/jasmine": "^3.5.4",
+    "@types/node": "^13.7.1",
+    "codelyzer": "^5.2.1",
     "core-js": "^2.5.7",
-    "jasmine-ts": "^0.2.1",
+    "jasmine": "^3.5.0",
+    "jasmine-ts": "^0.3.0",
     "karma": "^4.0.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-typescript": "^4.0.0",
-    "rxjs": "^6.3.3",
-    "ts-node": "^7.0.0",
-    "tslint": "^5.11.0",
-    "typescript": "^2.9.2",
-    "zone.js": "^0.8.26"
+    "prettier": "^1.19.1",
+    "rxjs": "^6.5.4",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.20.1",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "^3.7.5",
+    "zone.js": "^0.10.2"
   }
 }
