@@ -7,23 +7,25 @@ import { Shallow } from '../shallow';
 ////// Module Setup //////
 @Component({
   selector: 'service-response-label',
-  template: '<label>{{labelText}}</label>',
+  template: '<label>{{labelText}}</label>'
 })
 class FooLabelComponent {
-  labelText: string;
+  labelText?: string;
 
   constructor(httpClient: HttpClient) {
-    httpClient.get<string>('/foo/as/a/service').toPromise()
+    httpClient
+      .get<string>('/foo/as/a/service')
+      .toPromise()
       .then(
-        response => this.labelText = response,
-        () => this.labelText = 'ERROR'
+        response => (this.labelText = response),
+        () => (this.labelText = 'ERROR')
       );
   }
 }
 
 @NgModule({
   imports: [HttpClientModule],
-  declarations: [FooLabelComponent],
+  declarations: [FooLabelComponent]
 })
 class FooLabelModule {}
 //////////////////////////
@@ -32,13 +34,12 @@ describe('using replaceModule', () => {
   let shallow: Shallow<FooLabelComponent>;
 
   beforeEach(() => {
-    shallow = new Shallow(FooLabelComponent, FooLabelModule)
-      .replaceModule(HttpClientModule, HttpClientTestingModule);
+    shallow = new Shallow(FooLabelComponent, FooLabelModule).replaceModule(HttpClientModule, HttpClientTestingModule);
   });
 
   it('displays the response from the foo service', fakeAsync(async () => {
-    const {element, get, fixture} = await shallow.render();
-    const client = get(HttpTestingController as Type<HttpTestingController>);
+    const { element, inject, fixture } = await shallow.render();
+    const client = inject(HttpTestingController as Type<HttpTestingController>);
     client.expectOne('/foo/as/a/service').flush('foo response');
     tick();
     fixture.detectChanges();
@@ -47,8 +48,8 @@ describe('using replaceModule', () => {
   }));
 
   it('displays ERROR when a service error occurs', fakeAsync(async () => {
-    const {element, get, fixture} = await shallow.render();
-    const client = get(HttpTestingController as Type<HttpTestingController>);
+    const { element, inject, fixture } = await shallow.render();
+    const client = inject(HttpTestingController as Type<HttpTestingController>);
     client.expectOne('/foo/as/a/service').error(new ErrorEvent('BOOM'));
     tick();
     fixture.detectChanges();
