@@ -12,14 +12,14 @@ class RedService {
 @Component({
   selector: 'color-label',
   template: '<label>{{redService.color()}}</label>',
-  providers: [RedService], // Component has it's own provider
+  providers: [RedService] // Component has it's own provider
 })
 class ColorLabelComponent {
   constructor(public redService: RedService) {}
 }
 
 @NgModule({
-  declarations: [ColorLabelComponent],
+  declarations: [ColorLabelComponent]
   // The provider is not here, it's on the component
 })
 class ColorModule {}
@@ -29,12 +29,11 @@ describe('component with own provider', () => {
   let shallow: Shallow<ColorLabelComponent>;
 
   beforeEach(() => {
-    shallow = new Shallow(ColorLabelComponent, ColorModule)
-      .mock(RedService, {color: () => 'MOCKED COLOR'});
+    shallow = new Shallow(ColorLabelComponent, ColorModule).mock(RedService, { color: () => 'MOCKED COLOR' });
   });
 
   it('Uses the color from the RedService', async () => {
-    const {element} = await shallow.render();
+    const { element } = await shallow.render();
 
     expect(element.nativeElement.innerText).toBe('MOCKED COLOR');
   });
