@@ -21,10 +21,10 @@ class StepService {
     <div *ngFor="let step of stepService.getSteps()">
       <ng-template *ngComponentOutlet="step"></ng-template>
     </div>
-`,
+  `
 })
 class StepDisplayComponent {
-  constructor(public stepService: StepService) { }
+  constructor(public stepService: StepService) {}
 }
 
 @NgModule({
@@ -37,21 +37,21 @@ class StepModule {}
 
 describe('component with ngComponentOutlet and entry component', () => {
   let shallow: Shallow<StepDisplayComponent>;
-  @Component({selector: 'dummy-step-one', template: '<i></i>'})
+  @Component({ selector: 'dummy-step-one', template: '<i></i>' })
   class DummyStepOne {}
 
-  @Component({selector: 'dummy-step-two', template: '<i></i>'})
+  @Component({ selector: 'dummy-step-two', template: '<i></i>' })
   class DummyStepTwo {}
 
   beforeEach(() => {
     shallow = new Shallow(StepDisplayComponent, StepModule)
       .declare(DummyStepOne, DummyStepTwo)
       .dontMock(DummyStepOne, DummyStepTwo)
-      .mock(StepService, {getSteps: () => [DummyStepOne, DummyStepTwo]});
+      .mock(StepService, { getSteps: () => [DummyStepOne, DummyStepTwo] });
   });
 
   it('renders dynamic steps', async () => {
-    const {find} = await shallow.render();
+    const { find } = await shallow.render();
 
     expect(find(DummyStepOne)).toHaveFoundOne();
     expect(find(DummyStepTwo)).toHaveFoundOne();
