@@ -11,7 +11,7 @@ class RedService {
 
 @Component({
   selector: 'color-label',
-  template: '<label>{{redService.color()}}</label>',
+  template: '<label>{{redService.color()}}</label>'
 })
 class ColorLabelComponent {
   constructor(public redService: RedService) {}
@@ -19,7 +19,7 @@ class ColorLabelComponent {
 
 @NgModule({
   declarations: [ColorLabelComponent],
-  providers: [RedService],
+  providers: [RedService]
 })
 class ColorModule {}
 //////////////////////////
@@ -28,12 +28,11 @@ describe('component with service', () => {
   let shallow: Shallow<ColorLabelComponent>;
 
   beforeEach(() => {
-    shallow = new Shallow(ColorLabelComponent, ColorModule)
-      .mock(RedService, {color: () => 'MOCKED COLOR'});
+    shallow = new Shallow(ColorLabelComponent, ColorModule).mock(RedService, { color: () => 'MOCKED COLOR' });
   });
 
   it('Uses the color from the RedService', async () => {
-    const {element} = await shallow.render();
+    const { element } = await shallow.render();
 
     expect(element.nativeElement.innerText).toBe('MOCKED COLOR');
   });
