@@ -13,7 +13,7 @@ const STYLE_TOKEN = new InjectionToken<CustomStyles>('My custom styles');
     <label [class]="styles.defaultLabelClass">
       <ng-content></ng-content>
     </label>
-  `,
+  `
 })
 class LabelTextComponent {
   constructor(@Inject(STYLE_TOKEN) public styles: CustomStyles) {}
@@ -21,9 +21,7 @@ class LabelTextComponent {
 
 @NgModule({
   declarations: [LabelTextComponent],
-  providers: [
-    {provide: STYLE_TOKEN, useValue: {defaultLabelClass: 'uppercase font-size-small'}}
-  ]
+  providers: [{ provide: STYLE_TOKEN, useValue: { defaultLabelClass: 'uppercase font-size-small' } }]
 })
 class LabelTextModule {}
 //////////////////////////
@@ -32,19 +30,18 @@ describe('simple component example', () => {
   let shallow: Shallow<LabelTextComponent>;
 
   beforeEach(() => {
-    shallow = new Shallow(LabelTextComponent, LabelTextModule)
-      .mock(STYLE_TOKEN, {defaultLabelClass: 'MOCK-CLASS'});
+    shallow = new Shallow(LabelTextComponent, LabelTextModule).mock(STYLE_TOKEN, { defaultLabelClass: 'MOCK-CLASS' });
   });
 
   it('sets the color to the configured color and size', async () => {
-    const {find} = await shallow.render();
+    const { find } = await shallow.render();
 
     const label = find('label');
     expect(label.nativeElement.className).toContain('MOCK-CLASS');
   });
 
   it('places content in a label', async () => {
-    const {find} = await shallow.render('<label-text>Woot!</label-text>');
+    const { find } = await shallow.render('<label-text>Woot!</label-text>');
 
     const label = find('label');
     expect(label.nativeElement.innerText.trim()).toBe('Woot!');
