@@ -11,20 +11,21 @@ class RedService {
 
 @Component({
   selector: 'color-label',
-  template: '<label>{{redService.color()}}</label>',
+  template: '<label>{{redService.color()}}</label>'
 })
 class ColorLabelComponent {
   constructor(public redService: RedService) {}
 }
 
 @NgModule({
-  declarations: [ColorLabelComponent],
+  declarations: [ColorLabelComponent]
 })
-class ColorModule { // tslint:disable-line no-unnecessary-class
+class ColorModule {
+  // tslint:disable-line no-unnecessary-class
   static forRoot(): ModuleWithProviders {
     return {
       ngModule: ColorModule,
-      providers: [RedService],
+      providers: [RedService]
     };
   }
 }
@@ -34,12 +35,11 @@ describe('module with forRoot', () => {
   let shallow: Shallow<ColorLabelComponent>;
 
   beforeEach(() => {
-    shallow = new Shallow(ColorLabelComponent, ColorModule.forRoot())
-      .mock(RedService, {color: () => 'MOCKED COLOR'});
+    shallow = new Shallow(ColorLabelComponent, ColorModule.forRoot()).mock(RedService, { color: () => 'MOCKED COLOR' });
   });
 
   it('Uses the color from the RedService', async () => {
-    const {element} = await shallow.render();
+    const { element } = await shallow.render();
 
     expect(element.nativeElement.innerText).toBe('MOCKED COLOR');
   });
