@@ -4,25 +4,29 @@ import { Shallow } from '../shallow';
 ////// Module Setup //////
 @Component({
   selector: 'foo',
-  template: `<label>foo</label>`,
+  template: `
+    <label>foo</label>
+  `
 })
 class FooComponent {}
 
 @NgModule({
   declarations: [FooComponent],
-  exports: [FooComponent],
+  exports: [FooComponent]
 })
 class FooModule {}
 
 @Component({
   selector: 'bar',
-  template: `<label>bar</label>`,
+  template: `
+    <label>bar</label>
+  `
 })
 class BarComponent {}
 
 @NgModule({
   declarations: [BarComponent],
-  exports: [BarComponent],
+  exports: [BarComponent]
 })
 class BarModule {}
 
@@ -31,7 +35,7 @@ class BarModule {}
   template: `
     <foo *ngIf="showFoo"></foo>
     <bar *ngIf="showBar"></bar>
-  `,
+  `
 })
 class FooBarComponent {
   @Input() showFoo = true;
@@ -40,7 +44,7 @@ class FooBarComponent {
 
 @NgModule({
   imports: [FooModule, BarModule],
-  declarations: [FooBarComponent],
+  declarations: [FooBarComponent]
 })
 class FooBarModule {}
 
@@ -54,21 +58,21 @@ describe('multi module', () => {
   });
 
   it('renders a foo and a bar by default', async () => {
-    const {find} = await shallow.render(`<foo-bar></foo-bar>`);
+    const { find } = await shallow.render(`<foo-bar></foo-bar>`);
 
     expect(find('foo')).toHaveFound(1);
     expect(find('bar')).toHaveFound(1);
   });
 
   it('does not render a foo when showFoo is false', async () => {
-    const {find} = await shallow.render(`<foo-bar [showFoo]="false"></foo-bar>`);
+    const { find } = await shallow.render(`<foo-bar [showFoo]="false"></foo-bar>`);
 
     expect(find('foo')).toHaveFound(0);
     expect(find('bar')).toHaveFound(1);
   });
 
   it('does not render a bar when showBar is false', async () => {
-    const {find} = await shallow.render(`<foo-bar [showBar]="false"></foo-bar>`);
+    const { find } = await shallow.render(`<foo-bar [showBar]="false"></foo-bar>`);
 
     expect(find('foo')).toHaveFound(1);
     expect(find('bar')).toHaveFound(0);
