@@ -5,7 +5,9 @@ import { Shallow } from '../shallow';
 // This service is provided globally via a ServiceModule
 @Injectable()
 class RedService {
-  color() { return 'RED'; }
+  color() {
+    return 'RED';
+  }
 }
 
 @NgModule({ providers: [RedService] })
@@ -15,7 +17,7 @@ class ServiceModule {}
 ////// Module Setup //////
 @Component({
   selector: 'color-label',
-  template: '<label>{{redService.color()}}</label>',
+  template: '<label>{{redService.color()}}</label>'
 })
 class ColorLabelComponent {
   constructor(public redService: RedService) {}
@@ -28,8 +30,7 @@ class ColorModule {}
 //////////////////////////////////////////////////////////////////
 // Somewhere in your top-level test setup (maybe the karma shim?)
 // This will import (and mock!) the ServiceModule in ALL shallow test modules
-Shallow.alwaysImport(ServiceModule)
-       .neverMock(RedService); // Without this line, the RedService will be auto-mocked
+Shallow.alwaysImport(ServiceModule).neverMock(RedService); // Without this line, the RedService will be auto-mocked
 //////////////////////////////////////////////////////////////////
 
 describe('alwaysImport', () => {
@@ -40,16 +41,16 @@ describe('alwaysImport', () => {
   });
 
   it('Uses the color from the RedService', async () => {
-    const {element} = await shallow.render('<color-label></color-label>');
+    const { element } = await shallow.render('<color-label></color-label>');
 
     // Using the actual service response here (not mocked)
     expect(element.nativeElement.innerText).toBe('RED');
   });
 
   it('Uses the color from the mocked RedService', async () => {
-    const {element} = await shallow
+    const { element } = await shallow
       // User mocks always override things that are 'neverMocked'
-      .mock(RedService, {color: () => 'MOCKED VALUE'})
+      .mock(RedService, { color: () => 'MOCKED VALUE' })
       .render('<color-label></color-label>');
 
     expect(element.nativeElement.innerText).toBe('MOCKED VALUE');
