@@ -11,7 +11,7 @@ class RedService {
 
 @Component({
   selector: 'color-label',
-  template: '<label>{{redService.color()}}</label>',
+  template: '<label>{{redService.color()}}</label>'
 })
 class ColorLabelComponent {
   constructor(public redService: RedService) {}
@@ -19,7 +19,7 @@ class ColorLabelComponent {
 
 @NgModule({
   declarations: [ColorLabelComponent],
-  providers: [RedService],
+  providers: [RedService]
 })
 class ColorModule {}
 //////////////////////////
@@ -34,12 +34,14 @@ describe('using dontMock', () => {
   let shallow: Shallow<ColorLabelComponent>;
 
   beforeEach(() => {
-    shallow = new Shallow(ColorLabelComponent, ColorModule)
-      .provideMock({provide: RedService, useValue: new MockRedService()});
+    shallow = new Shallow(ColorLabelComponent, ColorModule).provideMock({
+      provide: RedService,
+      useValue: new MockRedService()
+    });
   });
 
   it('Uses the color from the MockRedService', async () => {
-    const {element} = await shallow.render();
+    const { element } = await shallow.render();
 
     // Using the mocked service response here
     expect(element.nativeElement.innerText).toBe('FAKE RESPONSE');
