@@ -1,31 +1,50 @@
-import { Component, Directive, EventEmitter, Input, NgModule, OnInit, Output, TemplateRef, ViewContainerRef } from '@angular/core';
-import { InvalidBindOnEntryComponentError, InvalidInputBindError, InvalidStaticPropertyMockError, Renderer } from './renderer';
+import {
+  Component,
+  Directive,
+  EventEmitter,
+  Input,
+  NgModule,
+  OnInit,
+  Output,
+  TemplateRef,
+  ViewContainerRef
+} from '@angular/core';
+import {
+  InvalidBindOnEntryComponentError,
+  InvalidInputBindError,
+  InvalidStaticPropertyMockError,
+  Renderer
+} from './renderer';
 import { TestSetup } from './test-setup';
 
-class TestUtility { // tslint:disable-line no-unnecessary-class
+class TestUtility {
+  // tslint:disable-line no-unnecessary-class
   static readonly staticNumber = 123;
-  static staticMethod() { return 'foo'; }
+  static staticMethod() {
+    return 'foo';
+  }
 }
 
 const staticObject = {
   staticNumber: 123,
-  staticMethod: () => 'foo',
+  staticMethod: () => 'foo'
 };
 
 @Component({
   selector: 'thing',
   template: `
-    <div>{{myInput}}</div>
-    <span>{{promiseResult}}</span>
+    <div>{{ myInput }}</div>
+    <span>{{ promiseResult }}</span>
   `
 })
 class TestComponent implements OnInit {
-  @Input('renamedInput') fooInput: string;
-  @Input() myInput: string;
-  myProperty: string;
+  // tslint:disable-next-line: no-input-rename
+  @Input('renamedInput') fooInput!: string;
+  @Input() myInput!: string;
+  myProperty!: string;
   @Output() myOutput = new EventEmitter<any>();
   emitterWithoutOutputDecorator = new EventEmitter<any>();
-  promiseResult: string;
+  promiseResult!: string;
 
   async ngOnInit() {
     this.promiseResult = await Promise.resolve('Promise Result');
@@ -47,21 +66,21 @@ describe('Renderer', () => {
   });
 
   it('mocks static CLASS methods from the test setup', async () => {
-    setup.staticMocks.set(TestUtility, {staticMethod: () => 'mocked foo'});
+    setup.staticMocks.set(TestUtility, { staticMethod: () => 'mocked foo' });
     await renderer.render();
 
     expect(TestUtility.staticMethod()).toBe('mocked foo');
   });
 
   it('mocks static OBJECT methods from the test setup', async () => {
-    setup.staticMocks.set(staticObject, {staticMethod: () => 'mocked foo'});
+    setup.staticMocks.set(staticObject, { staticMethod: () => 'mocked foo' });
     await renderer.render();
 
     expect(staticObject.staticMethod()).toBe('mocked foo');
   });
 
   it('mocks static CLASS properties throws an error', async () => {
-    setup.staticMocks.set(TestUtility, {staticNumber: 999});
+    setup.staticMocks.set(TestUtility, { staticNumber: 999 });
 
     try {
       await renderer.render();
@@ -72,7 +91,7 @@ describe('Renderer', () => {
   });
 
   it('mocks static OBJECT properties throws an error', async () => {
-    setup.staticMocks.set(staticObject, {staticNumber: 999});
+    setup.staticMocks.set(staticObject, { staticNumber: 999 });
 
     try {
       await renderer.render();
@@ -83,7 +102,7 @@ describe('Renderer', () => {
   });
 
   it('spys on output event emitters', async () => {
-    const {instance} = await renderer.render();
+    const { instance } = await renderer.render();
     instance.myOutput.emit('FOO');
 
     // Spys have a `calls` property on them. This is the only way I know
@@ -93,18 +112,17 @@ describe('Renderer', () => {
   });
 
   it('does not spy on event emitters that are not marked as @Output', async () => {
-    const {instance} = await renderer.render();
+    const { instance } = await renderer.render();
     instance.emitterWithoutOutputDecorator.emit('FOO');
 
     // Spys have a `calls` property on them. This is the only way I know
     // how to detect an existing spy.
-    expect((instance.emitterWithoutOutputDecorator.emit as jasmine.Spy).calls)
-      .not.toBeDefined();
+    expect((instance.emitterWithoutOutputDecorator.emit as jasmine.Spy).calls).not.toBeDefined();
   });
 
   describe('with only template', () => {
     it('wraps the rendering in a container', async () => {
-      const {fixture} = await renderer.render('<thing></thing>');
+      const { fixture } = await renderer.render('<thing></thing>');
 
       expect(fixture.debugElement.children[0].componentInstance instanceof TestComponent).toBe(true);
     });
@@ -112,23 +130,22 @@ describe('Renderer', () => {
 
   describe('with no arguments', () => {
     it('wraps the rendering in a container', async () => {
-      const {fixture} = await renderer.render();
+      const { fixture } = await renderer.render();
 
       expect(fixture.debugElement.children[0].componentInstance instanceof TestComponent).toBe(true);
     });
-
   });
 
   describe('with only renderOptions', () => {
     it('wraps the rendering in a container', async () => {
-      const {fixture} = await renderer.render({});
+      const { fixture } = await renderer.render({});
 
       expect(fixture.debugElement.children[0].componentInstance instanceof TestComponent).toBe(true);
     });
 
     it('binds through the wrapper to the component', async () => {
-      const {instance} = await renderer.render({
-        bind: {myInput: 'FOO'}
+      const { instance } = await renderer.render({
+        bind: { myInput: 'FOO' }
       });
 
       expect(instance.myInput).toBe('FOO');
@@ -136,14 +153,15 @@ describe('Renderer', () => {
 
     it('works on renamed @Input properties', async () => {
       await renderer.render({
-        bind: {fooInput: 'FOO'}
+        bind: { fooInput: 'FOO' }
       });
+      expect(true).toBe(true);
     });
 
-    it('throws an error when binding to a property that is not marked as an @Input', async (done) => {
+    it('throws an error when binding to a property that is not marked as an @Input', async done => {
       try {
         await renderer.render({
-          bind: {myProperty: 'FOO'}
+          bind: { myProperty: 'FOO' }
         });
         fail('Render should have thrown an error because the myProperty is not an @Input');
       } catch (e) {
@@ -153,22 +171,17 @@ describe('Renderer', () => {
     });
 
     describe('without a selector defined on the component', () => {
-      @Component({template: '<div>Without selector</div>'})
+      @Component({ template: '<div>Without selector</div>' })
       class TestComponentWithoutSelector {}
 
-      @NgModule({declarations: [TestComponentWithoutSelector]})
+      @NgModule({ declarations: [TestComponentWithoutSelector] })
       class NoSelectorModule {}
 
       it('should be able to render without a template specified', async () => {
-        const testSetup = new TestSetup(
-          TestComponentWithoutSelector,
-          NoSelectorModule,
-        );
+        const testSetup = new TestSetup(TestComponentWithoutSelector, NoSelectorModule);
         testSetup.dontMock.push(TestComponentWithoutSelector);
-        //tslint:disable-next-line:no-shadowed-variable
-        const renderer = new Renderer(testSetup);
 
-        const {element} = await renderer.render();
+        const { element } = await new Renderer(testSetup).render();
 
         expect(element).toBeTruthy();
       });
@@ -177,13 +190,13 @@ describe('Renderer', () => {
 
   describe('whenStable', () => {
     it('is awaited by default', async () => {
-      const {find} = await renderer.render();
+      const { find } = await renderer.render();
 
       expect(find('span').nativeElement.textContent).toBe('Promise Result');
     });
 
     it('is not awaited when disabled in options', async () => {
-      const {find} = await renderer.render({whenStable: false});
+      const { find } = await renderer.render({ whenStable: false });
 
       expect(find('span').nativeElement.textContent).toBe('');
     });
@@ -191,17 +204,17 @@ describe('Renderer', () => {
 
   describe('detectChanges', () => {
     it('is detected by default', async () => {
-      const {find} = await renderer.render({
-        bind: {myInput: 'FOO'}
+      const { find } = await renderer.render({
+        bind: { myInput: 'FOO' }
       });
 
       expect(find('div').nativeElement.textContent).toBe('FOO');
     });
 
     it('is not detected when disabled in options', async () => {
-      const {find} = await renderer.render({
+      const { find } = await renderer.render({
         detectChanges: false,
-        bind: {myInput: 'FOO'}
+        bind: { myInput: 'FOO' }
       });
 
       expect(find('div').nativeElement.textContent).toBe('');
@@ -209,11 +222,11 @@ describe('Renderer', () => {
   });
 
   describe('structural directives', () => {
-    @Directive({ selector: '[ifOdd]'})
+    @Directive({ selector: '[ifOdd]' })
     class IfOddDirective {
       private hasView = false;
 
-      constructor(private readonly _templateRef: TemplateRef<any>, private readonly _viewContainer: ViewContainerRef) { }
+      constructor(private readonly _templateRef: TemplateRef<any>, private readonly _viewContainer: ViewContainerRef) {}
 
       @Input() set ifOdd(possiblyOdd: number) {
         const isOdd = possiblyOdd % 2 !== 0;
@@ -226,26 +239,26 @@ describe('Renderer', () => {
         }
       }
     }
-    @NgModule({declarations: [IfOddDirective]})
+    @NgModule({ declarations: [IfOddDirective] })
     class OddModule {}
 
     it('element is the first child element when testing a structural directive', async () => {
       const myRenderer = new Renderer(new TestSetup(IfOddDirective, OddModule));
-      const {element} = await myRenderer.render('<b *ifOdd="2"></b>');
+      const { element } = await myRenderer.render('<b *ifOdd="2"></b>');
 
       expect(element.nativeElement.tagName).toBe('B');
     });
 
     it('element is undefined when the structural directive does not render an element', async () => {
       const myRenderer = new Renderer(new TestSetup(IfOddDirective, OddModule));
-      const {element} = await myRenderer.render('<b *ifOdd="1"></b>');
+      const { element } = await myRenderer.render('<b *ifOdd="1"></b>');
 
       expect(element).not.toBeDefined();
     });
 
     it('instance is the directive instance when testing a structural directive', async () => {
       const myRenderer = new Renderer(new TestSetup(IfOddDirective, OddModule));
-      const {instance} = await myRenderer.render('<b *ifOdd="2"></b>');
+      const { instance } = await myRenderer.render('<b *ifOdd="2"></b>');
 
       expect(instance instanceof IfOddDirective).toBe(true);
     });
@@ -256,7 +269,7 @@ describe('Renderer', () => {
       @Component({
         template: '<i class="my-entry">My Entry</i>'
       })
-      class EntryComponent { }
+      class EntryComponent {}
 
       @Component({
         selector: 'my-normal-component',
@@ -274,7 +287,7 @@ describe('Renderer', () => {
 
       const mySetup = new TestSetup(NormalComponent, EntryTestModule);
       mySetup.dontMock.push(NormalComponent, EntryComponent);
-      const {find} = await new Renderer(mySetup).render({whenStable: true});
+      const { find } = await new Renderer(mySetup).render({ whenStable: true });
 
       expect(find('.my-entry')).toHaveFoundOne();
     });
@@ -284,12 +297,12 @@ describe('Renderer', () => {
         selector: 'child-component',
         template: '<i class="my-child">My Dependency</i>'
       })
-      class ChildComponent { }
+      class ChildComponent {}
 
       @Component({
         template: '<i class="my-entry"><child-component></child-component></i>'
       })
-      class EntryComponent { }
+      class EntryComponent {}
 
       @Component({
         selector: 'normal-component',
@@ -307,7 +320,7 @@ describe('Renderer', () => {
 
       const mySetup = new TestSetup(NormalComponent, EntryTestModule);
       mySetup.dontMock.push(NormalComponent, EntryComponent);
-      const {find} = await new Renderer(mySetup).render({whenStable: true});
+      const { find } = await new Renderer(mySetup).render({ whenStable: true });
 
       expect(find('.my-entry')).toHaveFoundOne();
     });
@@ -317,7 +330,7 @@ describe('Renderer', () => {
         template: '<i class="my-entry">My Entry</i>'
       })
       class EntryComponent {
-        @Input() devMadeAMistakeAndCreatedAnInputOnAnEntryComponent: string;
+        @Input() devMadeAMistakeAndCreatedAnInputOnAnEntryComponent!: string;
       }
 
       @NgModule({
@@ -329,7 +342,7 @@ describe('Renderer', () => {
       const mySetup = new TestSetup(EntryComponent, EntryTestModule);
       try {
         await new Renderer(mySetup).render({
-          bind: {devMadeAMistakeAndCreatedAnInputOnAnEntryComponent: 'Whoops!'}
+          bind: { devMadeAMistakeAndCreatedAnInputOnAnEntryComponent: 'Whoops!' }
         });
         fail('Should not have rendered the entry component');
       } catch (e) {
