@@ -21,67 +21,93 @@ export class Rendering<TComponent, TBindings> {
     private readonly _setup: TestSetup<TComponent>
   ) {}
 
+  readonly outputs: PickByType<TComponent, EventEmitter<any>> = outputProxy(this.instance);
+
   /////////////////////////////////////////////////////////////////////////////
   // The following methods MUST be arrow functions so they can be deconstructured
   // off of the class
   /////////////////////////////////////////////////////////////////////////////
-  public find = (cssOrDirective: string | Type<any>, options?: {query?: string}): QueryMatch<DebugElement> => {
-    const query = typeof cssOrDirective === 'string'
-      ? By.css(cssOrDirective)
-      : By.directive(this._setup.mockCache.find(cssOrDirective) || cssOrDirective);
+  readonly find = (cssOrDirective: string | Type<any>, options?: { query?: string }): QueryMatch<DebugElement> => {
+    const query =
+      typeof cssOrDirective === 'string'
+        ? By.css(cssOrDirective)
+        : By.directive(this._setup.mockCache.find(cssOrDirective) || cssOrDirective);
     const mainQuery = this.fixture.debugElement.queryAll(query);
-    const found = options && options.query
-      ? this.fixture.debugElement.queryAll(By.css(options.query)).filter(item => mainQuery.includes(item))
-      : mainQuery;
+    const found =
+      options && options.query
+        ? this.fixture.debugElement.queryAll(By.css(options.query)).filter(item => mainQuery.includes(item))
+        : mainQuery;
 
     if (found.includes(this.element)) {
       throw new Error(`Don't search for your test component, it is automatically returned by the shallow renderer`);
     }
 
     return createQueryMatch(found);
-  }
+  };
 
-  readonly findComponent = <TMatch>(component: Type<TMatch>, options?: {query?: string}): QueryMatch<TMatch> =>
-    this.findDirective(component, options)
+  readonly findComponent = <TMatch>(component: Type<TMatch>, options?: { query?: string }): QueryMatch<TMatch> =>
+    this.findDirective(component, options);
 
-  readonly findDirective = <TDirective>(directive: Type<TDirective>, options?: {query?: string}): QueryMatch<TDirective> => {
+  readonly findDirective = <TDirective>(
+    directive: Type<TDirective>,
+    options?: { query?: string }
+  ): QueryMatch<TDirective> => {
     const directiveOrMock = this._setup.mockCache.find(directive) || directive;
-    const foundElements = options && options.query
-      ? this.fixture.debugElement.queryAll(By.css(options.query))
-      : this.find(directive, options);
+    const foundElements =
+      options && options.query
+        ? this.fixture.debugElement.queryAll(By.css(options.query))
+        : this.find(directive, options);
     const foundDirectives = foundElements
       .map(result => {
-        try { return result.injector.get<TDirective>(directiveOrMock); } catch (e) { return undefined; }
+        try {
+          return result.injector.get<TDirective>(directiveOrMock);
+        } catch (e) {
+          return undefined;
+        }
       })
       .filter(i => i) as TDirective[];
-    if (foundDirectives.some(i => i as any === this.instance)) {
+    if (foundDirectives.some(i => (i as any) === this.instance)) {
       throw new Error(`Don't search for your test component, it is automatically returned by the shallow renderer`);
     }
     return createQueryMatch(foundDirectives);
-  }
+  };
 
-  readonly get = <TClass>(queryClass: Type<TClass>): TClass => TestBed.get(queryClass);
+  /**
+   * @deprecated Use inject instead
+   */
+  readonly get = <TClass>(queryClass: Type<TClass>): TClass => TestBed.inject(queryClass);
 
-  readonly outputs: PickByType<TComponent, EventEmitter<any>> = outputProxy(this.instance);
+  readonly inject: typeof TestBed['inject'] = (...args: Parameters<typeof TestBed['inject']>) =>
+    TestBed.inject(...args);
 
-  readonly findStructuralDirective = <TDirective>(directiveClass: Type<TDirective>, options?: {query?(d: TDirective): boolean}) =>
+  readonly findStructuralDirective = <TDirective>(
+    directiveClass: Type<TDirective>,
+    options?: { query?(d: TDirective): boolean }
+  ) =>
     createQueryMatch(
-      this.fixture.debugElement.queryAllNodes(node => {
-        try {
-          const instance = node.injector.get(directiveClass);
-          if (instance) {
-            return options && options.query ? options.query(instance) : true;
-          }
-        } catch (e) { } // tslint:disable-line no-empty
-        return false;
-      })
-      .map(node => node.injector.get<TDirective>(directiveClass))
-    )
+      this.fixture.debugElement
+        .queryAllNodes(node => {
+          try {
+            const instance = node.injector.get(directiveClass);
+            if (instance) {
+              return options && options.query ? options.query(instance) : true;
+            }
+          } catch (e) {} // tslint:disable-line no-empty
+          return false;
+        })
+        .map(node => node.injector.get<TDirective>(directiveClass))
+    );
 
-  readonly renderStructuralDirective = (directiveClassOrObject: Type<any> | QueryMatch<any> | object, renderContents = true) => {
-    const directives: Array<MockedDirective<{}>> = typeof directiveClassOrObject === 'function'
-      ? this.findStructuralDirective<MockedDirective<{}>>(directiveClassOrObject)
-      : directiveClassOrObject.length ? directiveClassOrObject : [directiveClassOrObject];
+  readonly renderStructuralDirective = (
+    directiveClassOrObject: Type<any> | QueryMatch<any> | object,
+    renderContents = true
+  ) => {
+    const directives: Array<MockedDirective<{}>> =
+      typeof directiveClassOrObject === 'function'
+        ? this.findStructuralDirective<MockedDirective<{}>>(directiveClassOrObject)
+        : directiveClassOrObject.length
+        ? directiveClassOrObject
+        : [directiveClassOrObject];
 
     if (!directives.length) {
       throw new Error(`Tried to render a structural directive but none were found.`);
@@ -90,7 +116,9 @@ export class Rendering<TComponent, TBindings> {
     directives.forEach(foundDirective => {
       if (!foundDirective.__viewContainer) {
         const directiveName = Object.getPrototypeOf(foundDirective).constructor.name;
-        throw new Error(`You may only manually render mocked directives with "renderStructuralDirective". Tried to render a structural directive (${directiveName}) but the directive is not mocked.`);
+        throw new Error(
+          `You may only manually render mocked directives with "renderStructuralDirective". Tried to render a structural directive (${directiveName}) but the directive is not mocked.`
+        );
       }
       if (renderContents) {
         foundDirective.__render();
@@ -98,5 +126,5 @@ export class Rendering<TComponent, TBindings> {
         foundDirective.__viewContainer.clear();
       }
     });
-  }
+  };
 }
